Title: 합과 차의 나눗셈, Time: 148ms, Memory: 10MB, Status: Passed - CodeTree

+++ 240408/합과 차의 나눗셈/divide-of-sum-and-sub.java
@@ -7,7 +7,7 @@ public class Main {
         int a = sc.nextInt();
         int b = sc.nextInt();
 
-        double result = (a + b) / (a - b);
+        double result = (double)(a + b) / (a - b);
 
         System.out.printf("%.2f", result);
     }
